pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = "Проект для управления заказами"
authors = ["AruukeNurpaiaskyzy <aruukelisa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
graphene-django = "^3.2.2"
django-graphql-jwt = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
